requirements.txt
Adafruit-DHT==1.4.0
certifi==2019.11.28
chardet==3.0.4
idna==2.9
influxdb==5.2.3
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
